Web-Apps/1. English Dictionary - Flask/requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
playsound==1.3.0
requests==2.28.0
urllib3==1.26.9
Werkzeug==2.1.2
wordnik-py3==2.1.2
